feat(03-02): decompose data-service.ts into focused generator modules
Split 654-LOC monolith into focused modules:
- generators/base.ts: BASE_METRICS map + getBaseMetric helper
- generators/churn.ts, lead-attribution.ts, lead-trends.ts,
  growth.ts, year-over-year.ts, top-partners.ts: individual generators
- cache.ts: DataCacheService class
- service.ts: AnalyticsDataService class + singleton

Updated all consumer imports. All 133 tests pass.
